Higher half rewrite completed

+++ kernel/drivers/vga.c
@@ -1,12 +1,11 @@
 #include <stdbool.h>
-#include <stddef.h>
 #include <stdint.h>
 #include <string.h>
 
-#include <learnix/vga.h>
+#include <learnix/drivers/vga.h>
 
-static size_t terminal_row;
-static size_t terminal_column;
+static uint32_t terminal_row;
+static uint32_t terminal_column;
 static uint8_t terminal_color;
 static uint16_t* terminal_buffer;
 
@@ -15,9 +14,9 @@ void terminal_initialize(void) {
 	terminal_column = 0;
 	terminal_color = vga_entry_color(VGA_COLOR_LIGHT_GREY, VGA_COLOR_BLACK);
 	terminal_buffer = VGA_MEMORY;
-	for (size_t y = 0; y < VGA_HEIGHT; y++) {
-		for (size_t x = 0; x < VGA_WIDTH; x++) {
-			const size_t index = y * VGA_WIDTH + x;
+	for (uint32_t y = 0; y < VGA_HEIGHT; y++) {
+		for (uint32_t x = 0; x < VGA_WIDTH; x++) {
+			const uint32_t index = y * VGA_WIDTH + x;
 			terminal_buffer[index] = vga_entry(' ', terminal_color);
 		}
 	}
@@ -27,8 +26,8 @@ void terminal_setcolor(uint8_t color) {
 	terminal_color = color;
 }
 
-void terminal_putentryat(unsigned char c, uint8_t color, size_t x, size_t y) {
-	const size_t index = y * VGA_WIDTH + x;
+void terminal_putentryat(unsigned char c, uint8_t color, uint32_t x, uint32_t y) {
+	const uint32_t index = y * VGA_WIDTH + x;
 	terminal_buffer[index] = vga_entry(c, color);
 }
 
@@ -36,7 +35,7 @@ void terminal_scroll() {
 	terminal_row = VGA_HEIGHT - 1;	// set active row to the last one
 
 	// move all rows except the first up by 1
-	for(size_t y = 0; y < VGA_HEIGHT - 1; y++) {
+	for(uint32_t y = 0; y < VGA_HEIGHT - 1; y++) {
 		memcpy(
 			terminal_buffer + y * VGA_WIDTH,
 			terminal_buffer + (y + 1) * VGA_WIDTH,
@@ -45,8 +44,8 @@ void terminal_scroll() {
 	}
 
 	// clean last row
-	for(size_t x = 0; x < VGA_WIDTH; x++) {
-		size_t curr = terminal_row * VGA_WIDTH + x;
+	for(uint32_t x = 0; x < VGA_WIDTH; x++) {
+		uint32_t curr = terminal_row * VGA_WIDTH + x;
 		terminal_buffer[curr] = vga_entry(' ', terminal_color);
 	}
 }
@@ -81,8 +80,8 @@ void terminal_putchar(char c) {
 	}	
 }
 
-void terminal_write(const char* data, size_t size) {
-	for (size_t i = 0; i < size; i++)
+void terminal_write(const char* data, uint32_t size) {
+	for (uint32_t i = 0; i < size; i++)
 		terminal_putchar(data[i]);
 }
