+++ libc/string/memcpy.c
@@ -6,7 +6,7 @@
     - EDI points to the destination
     rep simply repeates the following instruction as specified in the ECX register
 */
-void* memcpy(void* restrict dst, const void* restrict src, size_t size) {
+void* memcpy(void* restrict dst, const void* restrict src, uint32_t size) {
     asm volatile (
         "rep movsb"
         : "+D" (dst), "+S" (src), "+c" (size)
