+++ libc/string/memset.c
@@ -3,7 +3,7 @@
 /*
  * stobs is an x86 specific-instruction that enables hardware-optimized memory filling
 */
-void *memset(void *ptr, int value, size_t num) {
+void *memset(void *ptr, int value, uint32_t num) {
     asm volatile("cld; rep stosb\n" ::"D"(ptr), "a"(value), "c"(num)
                : "cc", "memory");
   return ptr;
