+++ libc/string/strlen.c
@@ -1,7 +1,7 @@
 #include <string.h>
 
-size_t strlen(const char* str) {
-    size_t i;
+uint32_t strlen(const char* str) {
+    uint32_t i;
     for(i = 0; str[i] != '\0'; i++);
     return i;
 }
